Create connector benchmark test
