+++ benchmark/src/main/scala/benchmark/connector/ldbc/Batch.scala
@@ -12,8 +12,6 @@ import scala.compiletime.uninitialized
 
 import org.openjdk.jmh.annotations.*
 
-import cats.*
-
 import cats.effect.*
 import cats.effect.unsafe.implicits.global
 
@@ -24,15 +22,21 @@ import ldbc.connector.*
 @BenchmarkMode(Array(Mode.Throughput))
 @OutputTimeUnit(TimeUnit.SECONDS)
 @State(Scope.Benchmark)
-class Select:
+class Batch:
 
   given Tracer[IO] = Tracer.noop[IO]
 
   @volatile
   var connection: Resource[IO, LdbcConnection[IO]] = uninitialized
 
+  @volatile
+  var values: String = uninitialized
+
+  @volatile
+  var records: List[(Int, String)] = List.empty
+
   @Setup
-  def setupDataSource(): Unit =
+  def setup(): Unit =
     connection = Connection[IO](
       host     = "127.0.0.1",
       port     = 13306,
@@ -42,23 +46,27 @@ class Select:
       ssl      = SSL.Trusted
     )
 
+    values = (1 to len).map(_ => "(?, ?)").mkString(",")
+
+    records = (1 to len).map(num => (num, s"record$num")).toList
+
   @Param(Array("10", "100", "1000", "2000", "4000"))
   var len: Int = uninitialized
 
   @Benchmark
-  def selectN: List[(Long, String)] =
+  def batchN(): Unit =
     connection
       .use { conn =>
         for
-          statement <- conn.prepareStatement("SELECT * FROM ldbc_test LIMIT ?")
-          _         <- statement.setInt(1, len)
-          resultSet <- statement.executeQuery()
-          records <- Monad[IO].whileM[List, (Long, String)](resultSet.next()) {
-                       for
-                         c1 <- resultSet.getLong(1)
-                         c2 <- resultSet.getString(2)
-                       yield (c1, c2)
-                     }
-        yield records
+          statement <- conn.prepareStatement(s"INSERT INTO ldbc_test (c1, c2) VALUES (?, ?)")
+          _ <- records.foldLeft(IO.unit) {
+                 case (acc, (id, value)) =>
+                   acc *>
+                     statement.setInt(1, id) *>
+                     statement.setString(2, value) *>
+                     statement.addBatch()
+               }
+          _ <- statement.executeBatch()
+        yield ()
       }
       .unsafeRunSync()
